FERENGI is now robust to a PSF centering error.
FERENGI contains a function which determines the centre of the PSF
within a PSF image and, if necessary, shifts the array such that the
PSF is perfectly centred.  The centre is determined by fitting a 2d
Gaussian, however this can occasionally fail, returning -ve sigma
and/or centres that lie outside the array.  This failure occurs when
the 'redshifted' and resampled psflo is so small that it is
effectively a single pixel.  This would be rare when going from SDSS
to HST, as the HST camera has much finer sampling, but happens above
z~0.1 for SDSS to SDSS.  To solve this problem, the code now does a
rough test for an incorrect 2d Gaussian fit, and if such a situation
is detected instead uses the shifts determined from a simple, but
robust, 'centre-of-mass' centroid.  A warning message is printed to
alert the user, although this substitution should generally not
significantly affect the result.