Some small FERENGI tweaks
Added an option /COUNTSOUT to produce output images in counts,
i.e. the same as the input image and so with the correct zeropoints,
rather than counts/sec.

Added more useful error message when ROBUST_FIT fails, which appears
to be related to there being too little sky in the input image.